Removing Strainer docs/example.
Changing the docs to reflect the removal of Strainer in favour of a Rake
alternative.

Adding ChefSpec tests to Rakefile

Forgot to add the spec tests to Rakefile.  Also amended docs to reflect
this addition.

Ignoring ChefSpec .coverage directory

We don't need Git to track this directory which gets created by ChefSepc
code coverage.