Upgrade to Berkshelf 3.
When generating a new cookbook with Meez the Gemfile only specifies
Berkshelf without a version number.  A bundle install will pull down the
latest Berkshelf which is version 3.x which has slightly changed the
sources are defined in the Berksfile.  Berkshelf 3 is a decent step
forward so is good to use anyway, however the main problem without
upgrading is that a `berks install` will complain due to the old style
Berksfile.